+++ CHANGELOG.md
@@ -6,9 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Implement the `Drain::flush` method added in [slog-rs/slog#349]
 * Define minimum supported rust version.
 * Setup Github Actions.
 * Fixup some minor typos in CHANGELOG.md, including an incorrect release date for 2.8.0.
+
+[slog-rs/slog#349]: https://github.com/slog-rs/slog/pull/349
 
 ## 2.8.0 - 2023-08-26